+++ ECommerce/src/ECommerce/Models/Item.cs
@@ -2,7 +2,7 @@
 {
     public class Item
     {
-        public Item(Product product, int quantity, decimal totalItemAmount)
+        public Item(Product product, int quantity, double totalItemAmount)
         {
             Product = product;
             Quantity = quantity;
@@ -12,19 +12,19 @@
 
         public Product Product { get; private set; }
         public int Quantity { get; private set; }
-        public decimal TotalItemAmount { get; private set; }
-        public decimal TotalDiscountedItemAmount { get; private set; }
+        public double TotalItemAmount { get; private set; }
+        public double TotalDiscountedItemAmount { get; private set; }
 
 
         public void SetQuantity(int qty)
         {
             Quantity = qty;
         }
-        public void SetTotalItemAmount(decimal amount)
+        public void SetTotalItemAmount(double amount)
         {
             TotalItemAmount = amount;
         }
-        public void SetTotalDiscountedItemAmount(decimal discountedAmount)
+        public void SetTotalDiscountedItemAmount(double discountedAmount)
         {
             TotalDiscountedItemAmount = discountedAmount;
         }
